gen_wg_config.sh: Allow to pass to the script a custom private key
Private keys can also be generated from the website, in manual ways so
if one is already present we can just create configuration files
starting from it.

To make this possible we need to pass the private key or set it manually
to the .json file, and then we can use it to generate the public key and
check if is already present.